requirements.txt
beautifulsoup4
apache-airflow-providers-telegram